Replace the old "temp RValue" removal with a new more agressive optimization.
This is a much more aggressive attempt at removing temporary "rvalues" inside
the CopyForwarding pass. It is now handles basic switch CFGs and does some
memory disambiguation.

This is still insufficient to handle the case that caused the worst benchmark
regressions, which are caused by UnfoldSequence. The problem is, that executes
an escaping closure during the temporary rvalue's lifetime. The only way around
that is to fully analyze both sides of the copy down through all uses to prove
they are completely nonescaping.

Note, the original intention of the CopyForwarding pass was a very quick Onone
peephole that did not require escape analysis or alias analysis.