Version 3 of TempRValueOptimization.
This version successfully optimizes UnfoldSequence, which is the holy grail.

Unfortunately, there's a problem building the stdlib which seems to be a
use-after-free. Note that we need to remove copied from CopiedDefs when we
eliminate them, but it doesn't seem to be working properly.

Once that's fixed, we need to run the benchmarks.