[sil-analysis] Add FunctionAnalysisBase::{hasAnalysis,maybeGet}(SILFunction *F).
Today, if one wants to invalidate state relative to your own function analysis,
you have to use FunctionAnalysisBase::get() to get the analysis. The problem
here is that if the analysis does not exist yet, then you are actually creating
the analysis. This is an issue when one wants to perform an action on an
analysis only if the analysis has already been built. An example of such a
situation is when one is processing a delete notification. If one does not have
an analysis for a function, one should just do nothing.

I am going to use this to fix a delete notification problem in
EpilogueARCAnalysis.

rdar://33841629
(cherry picked from commit eb4d94f10b6f66388ce9448fc00d49ce9fdc4e7b)