Fix ambiguity between overlay inits and default inits for BNNS structs.
Most of these are resolvable by simply switching over to inits without
the previously-defaulted arguments, because the default args are all-
or-nothing. In one case that doesn't make sense, so instead I've provided
a dedicated static func to construct one class of object, and an init that
doesn't match the importer-created one for the more common case.

resolves rdar://problem/33880553