Sema: Fix crash when ternary operator is applied to an lvalue of IUO type
Note that the bug was in Sema, but I added a SILGen test, since
that's an easy way to ensure the AST is valid.

Unfortunately the AST verifier does not perform many checks when
we emit diagnostics; adding the test case to an existing Sema test
would not catch regressions.

Fixes <rdar://problem/34048306>.