[silgen] Add a convenience constructor that constructs an RValue from a single ManagedValue.
We already have a constructor similar to this, but it requires you to pass in
the CanType of the ManagedValue. This convenience constructor in contrast just
uses the ManagedValues swift rvalue type, so the caller doesn't have to do
anything.

I am going to use this when changing self in SILGenApply to use an
ArgumentSource (so I can in certain cases treat self as an lvalue).

rdar://33358110