[silgen] Rather than passing in initialOptions as an out parameter to emitArgumentsForNormalApply, just use a return value.
The reason to do this is that we are already always passing in
ApplyOptions::None for this value at all call sites. By using a reference to
create an out parameter, we make it look like something more complex is occuring
here. In contrast the result makes it clear to the reader that there is no state
from the caller before the callee that can affect this variable.

rdar://33358110