[silgen] Pass foreignSelf to emitArgumentsForNormalApply by value instead of by ref.
By passing it by ref, we make it seem like foreignSelf is either an out
parameter or an inout parameter. By changing the API of
emitArgumentsForNormalApply to take foreignSelf as a value, the API clearly
communicates that foreignSelf is a value type that is passed as an in parameter.

rdar://33358110