[silgen] Sink foreignError into applyFirstLevelCallee leaf functions.
Again, we were not using foreignError as an out parameter despite the convention
and always passed in .None. In the places, where the leaf function was not
setting the foreignError before passing it off to another routine, I changed the
code to just pass in .None so I can just eliminate the variable entirely.

The only place where I had to actually sink instead of delete was
emitNormalApply.

rdar://33358110