[silgen] Add a test for objc factory constructors.
While updating DI for ownership, I noticed that we were closing a borrow scope too early in the DI version of this test file. While trying to find why I missed it earlier, I noticed that (AFAICT) we don't have a specific SILGen test for this behavior.

This commit adds the missing test with the ownership verifier disabled. I am
going to fix it in a subsequent commit.

rdar://31521023