[stdlib] Remove the Sequence requirement from BinaryInteger.Words
Adding this requirement turned out to significantly degrade the compile
times.

This patch keeps the proper implementation of `Words` for all the
builtin integer types, but restores the `_word(at:)` for the stdlib
internal use.  At the same time the default implementation of
`word(at:)` is provided in terms of `words` property, meaining that
developers don't need to implement the underscored API in order to get
the `BinaryInteger` conformance, as long as `Words` conforms to
`Collection` with indicies that conform to `BinaryInteger` and `Element`
is equal to `UInt`, which seems to be a reasonable limitation,
expecially given that it is only slightly stricter than the constraint
being removed from the `Words` associated type.